+++ test/query_parameter_definition_test.rb
@@ -18,6 +18,21 @@ class QueryParameterDefinitionTest < Minitest::Test
     assert_equal 1, param.ui_priority
   end
 
+  def test_bad_sql_list
+    bad_sqls = [
+      "INSERT INTO users (user_name) VALUES('Someone')",
+      "UPDATE users SET user_name = 'Someone'",
+      "DELETE FROM users WHERE user_name = 'Someone'"
+    ]
+    bad_sqls.each do |sql|
+      assert_raises(ArgumentError) do
+        Crossbeams::Dataminer::QueryParameterDefinition.new('col',
+                                                            :control_type => :list,
+                                                            :list_def => sql)
+      end
+    end
+  end
+
   def test_ordered_sql_list
     param = Crossbeams::Dataminer::QueryParameterDefinition.new('col', :control_type => :list, :list_def => 'SELECT code FROM table ORDER BY code')
     assert param.list_is_ordered?